requirements.txt
PyMuPDF
python-pptx
python-docx
google-generativeai
streamlit
python-dotenv